Refactor: Improve sleep and beverage data aggregation and error handling
This commit introduces several refactorings focused on the sleep and beverage tracking features. It improves data aggregation logic for different time periods, enhances error handling with more specific user-facing messages, and introduces new UI state classes for daily sleep views.

### Key Changes:

-   **Error Handling and UI Messages (`SleepOverviewViewModel.kt`):**
    -   Added specific string resources (`sleep_overview_error_loading`, `sleep_overview_error_saving`) for displaying errors related to loading or saving sleep data.
    -   The `saveSleep` function now includes robust error handling with a `try-catch` block and checks for a valid `profileId` before attempting to save, displaying an error message to the user on failure.

-   **Data Aggregation and Type Safety:**
    -   **`BeverageOverviewViewModel.kt`:**
        -   Refactored the data aggregation logic for weekly, monthly, and yearly views. The logic now first aggregates beverages by day and then calculates the average consumption for the larger period, ensuring more accurate calculations.
        -   The `amountMl` in `BeverageOverview.kt` is now a `Double` to accurately represent averaged values.
    -   **`SleepOverviewViewModel.kt`:**
        -   The `targetSleeps` flow now provides a non-nullable `UInt` with a default value, simplifying downstream logic.
        -   Removed commented-out and redundant code related to data flow mapping.

-   **New UI State for Daily Sleep View:**
    -   Introduced `SleepDailyUiState.kt` and `SleepDailyError.kt` to manage the state for the upcoming daily sleep details screen.
    -   The `SleepDailyViewModel` has been set up, inheriting from the base `DailyViewModel` and initialized with necessary dependencies, although the implementation is not yet complete.

-   **Minor Refinements:**
    -   In `SleepOverview.kt`, derived properties (`minutes`, `hours`) were removed to simplify the data class.
    -   Updated string resource files (`strings.xml`) in both default and German locales with the new error messages.

+++ composeApp/src/commonMain/kotlin/org/darthacheron/fitbe/health/beverages/BeverageOverview.kt
@@ -5,5 +5,5 @@ import org.darthacheron.fitbe.health.components.Overview
 
 data class BeverageOverview(
     override val date: LocalDate,
-    val amountMl: UInt
+    val amountMl: Double
 ) : Overview
